Consolidate inference cli and tuned model class
